@@ -1,10 +1,10 @@
 package awsbase
 
 import (
+	"bytes"
 	"context"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"net"
 	"net/http"
 	"os"
@@ -29,6 +29,7 @@ import (
 	"github.com/hashicorp/aws-sdk-go-base/v2/mockdata"
 	"github.com/hashicorp/aws-sdk-go-base/v2/servicemocks"
 	"github.com/hashicorp/aws-sdk-go-base/v2/useragent"
+	"github.com/hashicorp/terraform-plugin-log/tflogtest"
 )
 
 const (
@@ -975,14 +976,14 @@ aws_secret_access_key = DefaultSharedCredentialsSecretKey
 			}
 
 			if testCase.EnableWebIdentityEnvVars || testCase.EnableWebIdentityConfig {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-web-identity-token-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-web-identity-token-file")
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary web identity token file: %s", err)
 				}
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(file.Name(), []byte(servicemocks.MockWebIdentityToken), 0600)
+				err = os.WriteFile(file.Name(), []byte(servicemocks.MockWebIdentityToken), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing web identity token file: %s", err)
@@ -1007,7 +1008,7 @@ aws_secret_access_key = DefaultSharedCredentialsSecretKey
 			testCase.Config.StsEndpoint = stsEndpoint
 
 			if testCase.SharedConfigurationFile != "" {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -1015,7 +1016,7 @@ aws_secret_access_key = DefaultSharedCredentialsSecretKey
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
+				err = os.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing shared configuration file: %s", err)
@@ -1025,7 +1026,7 @@ aws_secret_access_key = DefaultSharedCredentialsSecretKey
 			}
 
 			if testCase.SharedCredentialsFile != "" {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-credentials-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-credentials-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared credentials file: %s", err)
@@ -1033,7 +1034,7 @@ aws_secret_access_key = DefaultSharedCredentialsSecretKey
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(file.Name(), []byte(testCase.SharedCredentialsFile), 0600)
+				err = os.WriteFile(file.Name(), []byte(testCase.SharedCredentialsFile), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing shared credentials file: %s", err)
@@ -1049,7 +1050,7 @@ aws_secret_access_key = DefaultSharedCredentialsSecretKey
 				os.Setenv(k, v)
 			}
 
-			awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
+			ctx, awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
 
 			if err != nil {
 				if testCase.ExpectedError == nil {
@@ -1068,7 +1069,7 @@ aws_secret_access_key = DefaultSharedCredentialsSecretKey
 				t.Fatalf("expected error, got no error")
 			}
 
-			credentialsValue, err := awsConfig.Credentials.Retrieve(context.Background())
+			credentialsValue, err := awsConfig.Credentials.Retrieve(ctx)
 
 			if err != nil {
 				t.Fatalf("unexpected credentials Retrieve() error: %s", err)
@@ -1101,14 +1102,12 @@ func testUserAgentProducts(t *testing.T, testCase test.UserAgentTestCase) {
 		httpSdkAgent = request.Header.Get("X-Amz-User-Agent")
 	})
 
-	awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
+	ctx, awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
 	if err != nil {
 		t.Fatalf("error in GetAwsConfig() '%[1]T': %[1]s", err)
 	}
 
-	client := stsClient(awsConfig, testCase.Config)
-
-	ctx := context.Background()
+	client := stsClient(ctx, awsConfig, testCase.Config)
 
 	if testCase.Context != nil {
 		ctx = useragent.Context(ctx, testCase.Context)
@@ -1342,7 +1341,7 @@ region = us-west-2
 			}
 
 			if testCase.SharedConfigurationFile != "" {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -1350,7 +1349,7 @@ region = us-west-2
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
+				err = os.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing shared configuration file: %s", err)
@@ -1361,7 +1360,7 @@ region = us-west-2
 
 			testCase.Config.SkipCredsValidation = true
 
-			awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
+			_, awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
 			if err != nil {
 				t.Fatalf("error in GetAwsConfig() '%[1]T': %[1]s", err)
 			}
@@ -1460,7 +1459,7 @@ max_attempts = 10
 			}
 
 			if testCase.SharedConfigurationFile != "" {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -1468,7 +1467,7 @@ max_attempts = 10
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
+				err = os.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing shared configuration file: %s", err)
@@ -1479,7 +1478,7 @@ max_attempts = 10
 
 			testCase.Config.SkipCredsValidation = true
 
-			awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
+			_, awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
 			if err != nil {
 				t.Fatalf("error in GetAwsConfig() '%[1]T': %[1]s", err)
 			}
@@ -1674,7 +1673,7 @@ use_fips_endpoint = true
 			}
 
 			if testCase.SharedConfigurationFile != "" {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -1682,7 +1681,7 @@ use_fips_endpoint = true
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
+				err = os.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing shared configuration file: %s", err)
@@ -1693,12 +1692,12 @@ use_fips_endpoint = true
 
 			testCase.Config.SkipCredsValidation = true
 
-			awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
+			ctx, awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
 			if err != nil {
 				t.Fatalf("error in GetAwsConfig() '%[1]T': %[1]s", err)
 			}
 
-			useFIPSState, _, err := awsconfig.ResolveUseFIPSEndpoint(context.Background(), awsConfig.ConfigSources)
+			useFIPSState, _, err := awsconfig.ResolveUseFIPSEndpoint(ctx, awsConfig.ConfigSources)
 			if err != nil {
 				t.Fatalf("error in ResolveUseFIPSEndpoint: %s", err)
 			}
@@ -1706,7 +1705,7 @@ use_fips_endpoint = true
 				t.Errorf("expected UseFIPSEndpoint %q, got: %q", awsconfig.FIPSEndpointStateString(e), awsconfig.FIPSEndpointStateString(a))
 			}
 
-			useDualStackState, _, err := awsconfig.ResolveUseDualStackEndpoint(context.Background(), awsConfig.ConfigSources)
+			useDualStackState, _, err := awsconfig.ResolveUseDualStackEndpoint(ctx, awsConfig.ConfigSources)
 			if err != nil {
 				t.Fatalf("error in ResolveUseDualStackEndpoint: %s", err)
 			}
@@ -1806,7 +1805,7 @@ func TestEC2MetadataServiceClientEnableState(t *testing.T) {
 			}
 
 			if testCase.SharedConfigurationFile != "" {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -1814,7 +1813,7 @@ func TestEC2MetadataServiceClientEnableState(t *testing.T) {
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
+				err = os.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing shared configuration file: %s", err)
@@ -1825,7 +1824,7 @@ func TestEC2MetadataServiceClientEnableState(t *testing.T) {
 
 			testCase.Config.SkipCredsValidation = true
 
-			awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
+			_, awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
 			if err != nil {
 				t.Fatalf("error in GetAwsConfig() '%[1]T': %[1]s", err)
 			}
@@ -1974,7 +1973,7 @@ ec2_metadata_service_endpoint = https://127.1.1.1:1111
 			}
 
 			if testCase.SharedConfigurationFile != "" {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -1982,7 +1981,7 @@ ec2_metadata_service_endpoint = https://127.1.1.1:1111
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
+				err = os.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing shared configuration file: %s", err)
@@ -1993,7 +1992,7 @@ ec2_metadata_service_endpoint = https://127.1.1.1:1111
 
 			testCase.Config.SkipCredsValidation = true
 
-			awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
+			_, awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
 			if err != nil {
 				t.Fatalf("error in GetAwsConfig() '%[1]T': %[1]s", err)
 			}
@@ -2096,7 +2095,7 @@ ec2_metadata_service_endpoint_mode = IPv4
 			}
 
 			if testCase.SharedConfigurationFile != "" {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -2104,7 +2103,7 @@ ec2_metadata_service_endpoint_mode = IPv4
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
+				err = os.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing shared configuration file: %s", err)
@@ -2115,7 +2114,7 @@ ec2_metadata_service_endpoint_mode = IPv4
 
 			testCase.Config.SkipCredsValidation = true
 
-			awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
+			_, awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
 			if err != nil {
 				t.Fatalf("error in GetAwsConfig() '%[1]T': %[1]s", err)
 			}
@@ -2221,7 +2220,7 @@ func TestCustomCABundle(t *testing.T) {
 				os.Setenv(k, v)
 			}
 
-			tempdir, err := ioutil.TempDir("", "temp")
+			tempdir, err := os.MkdirTemp("", "temp")
 			if err != nil {
 				t.Fatalf("error creating temp dir: %s", err)
 			}
@@ -2254,7 +2253,7 @@ func TestCustomCABundle(t *testing.T) {
 			}
 
 			if testCase.SetSharedConfigurationFile {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -2262,7 +2261,7 @@ func TestCustomCABundle(t *testing.T) {
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(
+				err = os.WriteFile(
 					file.Name(),
 					[]byte(fmt.Sprintf(`
 [default]
@@ -2278,7 +2277,7 @@ ca_bundle = %s
 			}
 
 			if testCase.SetSharedConfigurationFileToInvalid {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -2286,7 +2285,7 @@ ca_bundle = %s
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(
+				err = os.WriteFile(
 					file.Name(),
 					[]byte(`
 [default]
@@ -2303,7 +2302,7 @@ ca_bundle = no-such-file
 
 			testCase.Config.SkipCredsValidation = true
 
-			awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
+			_, awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
 			if err != nil {
 				t.Fatalf("error in GetAwsConfig() '%[1]T': %[1]s", err)
 			}
@@ -2445,7 +2444,7 @@ aws_secret_access_key = SharedConfigurationSourceSecretKey
 
 			testCase.Config.StsEndpoint = stsEndpoint
 
-			tempdir, err := ioutil.TempDir("", "temp")
+			tempdir, err := os.MkdirTemp("", "temp")
 			if err != nil {
 				t.Fatalf("error creating temp dir: %s", err)
 			}
@@ -2453,7 +2452,7 @@ aws_secret_access_key = SharedConfigurationSourceSecretKey
 			os.Setenv("TMPDIR", tempdir)
 
 			if testCase.SharedConfigurationFile != "" {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -2461,7 +2460,7 @@ aws_secret_access_key = SharedConfigurationSourceSecretKey
 
 				defer os.Remove(file.Name())
 
-				err = ioutil.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
+				err = os.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing shared configuration file: %s", err)
@@ -2472,7 +2471,7 @@ aws_secret_access_key = SharedConfigurationSourceSecretKey
 
 			testCase.Config.SkipCredsValidation = true
 
-			awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
+			ctx, awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
 
 			if err != nil {
 				if testCase.ExpectedError == nil {
@@ -2487,7 +2486,7 @@ aws_secret_access_key = SharedConfigurationSourceSecretKey
 				return
 			}
 
-			credentialsValue, err := awsConfig.Credentials.Retrieve(context.Background())
+			credentialsValue, err := awsConfig.Credentials.Retrieve(ctx)
 
 			if err != nil {
 				t.Fatalf("unexpected credentials Retrieve() error: %s", err)
@@ -2690,14 +2689,14 @@ web_identity_token_file = no-such-file
 
 			testCase.Config.StsEndpoint = stsEndpoint
 
-			tempdir, err := ioutil.TempDir("", "temp")
+			tempdir, err := os.MkdirTemp("", "temp")
 			if err != nil {
 				t.Fatalf("error creating temp dir: %s", err)
 			}
 			defer os.Remove(tempdir)
 			os.Setenv("TMPDIR", tempdir)
 
-			tokenFile, err := ioutil.TempFile("", "aws-sdk-go-base-web-identity-token-file")
+			tokenFile, err := os.CreateTemp("", "aws-sdk-go-base-web-identity-token-file")
 			if err != nil {
 				t.Fatalf("unexpected error creating temporary web identity token file: %s", err)
 			}
@@ -2705,7 +2704,7 @@ web_identity_token_file = no-such-file
 
 			defer os.Remove(tokenFileName)
 
-			err = ioutil.WriteFile(tokenFileName, []byte(servicemocks.MockWebIdentityToken), 0600)
+			err = os.WriteFile(tokenFileName, []byte(servicemocks.MockWebIdentityToken), 0600)
 
 			if err != nil {
 				t.Fatalf("unexpected error writing web identity token file: %s", err)
@@ -2731,7 +2730,7 @@ web_identity_token_file = no-such-file
 			}
 
 			if testCase.SharedConfigurationFile != "" {
-				file, err := ioutil.TempFile("", "aws-sdk-go-base-shared-configuration-file")
+				file, err := os.CreateTemp("", "aws-sdk-go-base-shared-configuration-file")
 
 				if err != nil {
 					t.Fatalf("unexpected error creating temporary shared configuration file: %s", err)
@@ -2743,7 +2742,7 @@ web_identity_token_file = no-such-file
 					testCase.SharedConfigurationFile += fmt.Sprintf("web_identity_token_file = %s\n", tokenFileName)
 				}
 
-				err = ioutil.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
+				err = os.WriteFile(file.Name(), []byte(testCase.SharedConfigurationFile), 0600)
 
 				if err != nil {
 					t.Fatalf("unexpected error writing shared configuration file: %s", err)
@@ -2754,7 +2753,7 @@ web_identity_token_file = no-such-file
 
 			testCase.Config.SkipCredsValidation = true
 
-			awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
+			ctx, awsConfig, err := GetAwsConfig(context.Background(), testCase.Config)
 
 			if err != nil {
 				if testCase.ExpectedError == nil {
@@ -2769,7 +2768,7 @@ web_identity_token_file = no-such-file
 				return
 			}
 
-			credentialsValue, err := awsConfig.Credentials.Retrieve(context.Background())
+			credentialsValue, err := awsConfig.Credentials.Retrieve(ctx)
 
 			if err != nil {
 				t.Fatalf("unexpected credentials Retrieve() error: %s", err)
@@ -2855,11 +2854,11 @@ func TestGetAwsConfigWithAccountIDAndPartition(t *testing.T) {
 			defer ts.Close()
 			tc.config.StsEndpoint = ts.URL
 
-			awsConfig, err := GetAwsConfig(context.Background(), tc.config)
+			ctx, awsConfig, err := GetAwsConfig(context.Background(), tc.config)
 			if err != nil {
 				t.Fatalf("expected no error from GetAwsConfig(), got: %s", err)
 			}
-			acctID, part, err := GetAwsAccountIDAndPartition(context.Background(), awsConfig, tc.config)
+			acctID, part, err := GetAwsAccountIDAndPartition(ctx, awsConfig, tc.config)
 			if err != nil {
 				if !tc.expectError {
 					t.Fatalf("expected no error, got: %s", err)
@@ -3059,7 +3058,7 @@ func TestRetryHandlers(t *testing.T) {
 				SecretKey:           servicemocks.MockStaticSecretKey,
 				SkipCredsValidation: true,
 			}
-			awsConfig, err := GetAwsConfig(context.Background(), config)
+			ctx, awsConfig, err := GetAwsConfig(context.Background(), config)
 			if err != nil {
 				t.Fatalf("unexpected error from GetAwsConfig(): %s", err)
 			}
@@ -3072,7 +3071,7 @@ func TestRetryHandlers(t *testing.T) {
 			}, func(i interface{}) interface{} {
 				return i
 			})
-			_, metadata, err := am.HandleFinalize(context.Background(), middleware.FinalizeInput{Request: nil}, testcase.NextHandler())
+			_, metadata, err := am.HandleFinalize(ctx, middleware.FinalizeInput{Request: nil}, testcase.NextHandler())
 			if err != nil && testcase.Err == nil {
 				t.Errorf("expect no error, got %v", err)
 			} else if err == nil && testcase.Err != nil {
@@ -3112,4 +3111,58 @@ func (r *withNoDelay) RetryDelay(attempt int, err error) (time.Duration, error) 
 	}
 
 	return 0 * time.Second, nil
+}
+
+func TestLogger(t *testing.T) {
+	var buf bytes.Buffer
+	ctx := tflogtest.RootLogger(context.Background(), &buf)
+
+	oldEnv := servicemocks.InitSessionTestEnv()
+	defer servicemocks.PopEnv(oldEnv)
+
+	config := &Config{
+		AccessKey: servicemocks.MockStaticAccessKey,
+		Region:    "us-east-1",
+		SecretKey: servicemocks.MockStaticSecretKey,
+	}
+
+	ts := servicemocks.MockAwsApiServer("STS", []*servicemocks.MockEndpoint{
+		servicemocks.MockStsGetCallerIdentityValidEndpoint,
+	})
+	defer ts.Close()
+	config.StsEndpoint = ts.URL
+
+	expectedName := fmt.Sprintf("provider.%s", loggerName)
+
+	ctx, awsConfig, err := GetAwsConfig(ctx, config)
+	if err != nil {
+		t.Fatalf("GetAwsConfig: unexpected '%[1]T': %[1]s", err)
+	}
+
+	lines, err := tflogtest.MultilineJSONDecode(&buf)
+	if err != nil {
+		t.Fatalf("GetAwsConfig: decoding log lines: %s", err)
+	}
+
+	for i, line := range lines {
+		if a, e := line["@module"], expectedName; a != e {
+			t.Errorf("GetAwsConfig: line %d: expected module %q, got %q", i+1, e, a)
+		}
+	}
+
+	_, _, err = GetAwsAccountIDAndPartition(ctx, awsConfig, config)
+	if err != nil {
+		t.Fatalf("GetAwsAccountIDAndPartition: unexpected '%[1]T': %[1]s", err)
+	}
+
+	lines, err = tflogtest.MultilineJSONDecode(&buf)
+	if err != nil {
+		t.Fatalf("GetAwsAccountIDAndPartition: decoding log lines: %s", err)
+	}
+
+	for i, line := range lines {
+		if a, e := line["@module"], expectedName; a != e {
+			t.Errorf("GetAwsAccountIDAndPartition: line %d: expected module %q, got %q", i+1, e, a)
+		}
+	}
 }
