@@ -4,19 +4,20 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"log"
 	"os"
-	"strings"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/credentials"
 	"github.com/aws/aws-sdk-go-v2/credentials/stscreds"
 	"github.com/aws/aws-sdk-go-v2/service/sts/types"
+	"github.com/hashicorp/aws-sdk-go-base/v2/logging"
 )
 
 func getCredentialsProvider(ctx context.Context, c *Config) (aws.CredentialsProvider, string, error) {
-	loadOptions, err := commonLoadOptions(c)
+	logger := logging.RetrieveLogger(ctx)
+
+	loadOptions, err := commonLoadOptions(ctx, c)
 	if err != nil {
 		return nil, "", err
 	}
@@ -28,7 +29,7 @@ func getCredentialsProvider(ctx context.Context, c *Config) (aws.CredentialsProv
 		}),
 		// The endpoint resolver is added here instead of in commonLoadOptions() so that it
 		// is not included in the aws.Config returned to the caller
-		config.WithEndpointResolverWithOptions(credentialsEndpointResolver(c)),
+		config.WithEndpointResolverWithOptions(credentialsEndpointResolver(ctx, c)),
 	)
 
 	envConfig, err := config.NewEnvConfig()
@@ -37,7 +38,7 @@ func getCredentialsProvider(ctx context.Context, c *Config) (aws.CredentialsProv
 	}
 
 	if c.Profile != "" && os.Getenv("AWS_ACCESS_KEY_ID") != "" && os.Getenv("AWS_SECRET_ACCESS_KEY") != "" {
-		log.Printf(`[WARN] A Profile was specified along with the environment variables "AWS_ACCESS_KEY_ID" and "AWS_SECRET_ACCESS_KEY". ` +
+		logger.Warn(ctx, `A Profile was specified along with the environment variables "AWS_ACCESS_KEY_ID" and "AWS_SECRET_ACCESS_KEY". `+
 			"The Profile is now used instead of the environment variable credentials. This may lead to unexpected behavior.")
 	}
 
@@ -58,11 +59,15 @@ func getCredentialsProvider(ctx context.Context, c *Config) (aws.CredentialsProv
 			for i, v := range sharedCredentialsFiles {
 				f[i] = fmt.Sprintf(`"%s"`, v)
 			}
-			log.Printf("[DEBUG] Using shared credentials files from configuration: [%s]", strings.Join(f, ", "))
+			logger.Debug(ctx, "Using shared credentials files from configuration", map[string]any{
+				"tf_aws.shared_credentials_files": f,
+			})
 		} else {
 			if envConfig.SharedCredentialsFile != "" {
-				log.Printf("[DEBUG] Using shared credentials file environment variables: %q", envConfig.SharedCredentialsFile)
 				sharedCredentialsFiles = []string{envConfig.SharedCredentialsFile}
+				logger.Debug(ctx, "Using shared credentials file environment variables", map[string]any{
+					"tf_aws.shared_credentials_files": sharedCredentialsFiles,
+				})
 			}
 		}
 
@@ -75,11 +80,15 @@ func getCredentialsProvider(ctx context.Context, c *Config) (aws.CredentialsProv
 			for i, v := range sharedConfigFiles {
 				f[i] = fmt.Sprintf(`"%s"`, v)
 			}
-			log.Printf("[DEBUG] Using shared configuration files from configuration: %v", strings.Join(f, ", "))
+			logger.Debug(ctx, "Using shared configuration files from configuration", map[string]any{
+				"tf_aws.shared_config_files": f,
+			})
 		} else {
 			if envConfig.SharedConfigFile != "" {
-				log.Printf("[DEBUG] Using shared configuration file environment variables: %s", envConfig.SharedConfigFile)
 				sharedConfigFiles = []string{envConfig.SharedConfigFile}
+				logger.Debug(ctx, "Using shared configuration file environment variables", map[string]any{
+					"tf_aws.shared_config_files": sharedConfigFiles,
+				})
 			}
 		}
 
@@ -98,7 +107,9 @@ func getCredentialsProvider(ctx context.Context, c *Config) (aws.CredentialsProv
 	// We need to validate both the configured and envvar named profiles for validity,
 	// but to use proper precedence, we only set the configured named profile
 	if c.Profile != "" {
-		log.Printf("[DEBUG] Using profile from configuration: %q", c.Profile)
+		logger.Debug(ctx, "Using profile from configuration", map[string]any{
+			"tf_aws.profile": c.Profile,
+		})
 		loadOptions = append(
 			loadOptions,
 			config.WithSharedConfigProfile(c.Profile),
@@ -116,7 +127,9 @@ func getCredentialsProvider(ctx context.Context, c *Config) (aws.CredentialsProv
 		if c.Token != "" {
 			params = append(params, "token")
 		}
-		log.Printf("[DEBUG] Using %s from configuration", strings.Join(params, ", "))
+		logger.Debug(ctx, "Using authentication parameters from configuration", map[string]any{
+			"tf_aws.auth_fields": params,
+		})
 		loadOptions = append(
 			loadOptions,
 			config.WithCredentialsProvider(
@@ -164,7 +177,9 @@ Error: %w`, err)
 		return cfg.Credentials, creds.Source, nil
 	}
 
-	log.Printf("[INFO] Retrieved initial credentials from %q", creds.Source)
+	logger.Info(ctx, "Retrieved initial credentials", map[string]any{
+		"tf_aws.credentials_source": creds.Source,
+	})
 	provider, err := assumeRoleCredentialsProvider(ctx, cfg, c)
 
 	return provider, creds.Source, err
@@ -172,7 +187,7 @@ Error: %w`, err)
 
 func webIdentityCredentialsProvider(ctx context.Context, awsConfig aws.Config, c *Config) (aws.CredentialsProvider, error) {
 	ar := c.AssumeRoleWithWebIdentity
-	client := stsClient(awsConfig, c)
+	client := stsClient(ctx, awsConfig, c)
 
 	appCreds := stscreds.NewWebIdentityRoleProvider(client, ar.RoleARN, ar, func(opts *stscreds.WebIdentityRoleOptions) {
 		opts.RoleSessionName = ar.SessionName
@@ -195,6 +210,8 @@ func webIdentityCredentialsProvider(ctx context.Context, awsConfig aws.Config, c
 }
 
 func assumeRoleCredentialsProvider(ctx context.Context, awsConfig aws.Config, c *Config) (aws.CredentialsProvider, error) {
+	logger := logging.RetrieveLogger(ctx)
+
 	ar := c.AssumeRole
 
 	if ar.RoleARN == "" {
@@ -202,9 +219,14 @@ func assumeRoleCredentialsProvider(ctx context.Context, awsConfig aws.Config, c 
 	}
 
 	// When assuming a role, we need to first authenticate the base credentials above, then assume the desired role
-	log.Printf("[INFO] Assuming IAM Role %q (SessionName: %q, ExternalId: %q, SourceIdentity: %q)", ar.RoleARN, ar.SessionName, ar.ExternalID, ar.SourceIdentity)
+	logger.Info(ctx, "Assuming IAM Role", map[string]any{
+		"tf_aws.assume_role.role_arn":        ar.RoleARN,
+		"tf_aws.assume_role.session_name":    ar.SessionName,
+		"tf_aws.assume_role.external_id":     ar.ExternalID,
+		"tf_aws.assume_role.source_identity": ar.SourceIdentity,
+	})
 
-	client := stsClient(awsConfig, c)
+	client := stsClient(ctx, awsConfig, c)
 
 	appCreds := stscreds.NewAssumeRoleProvider(client, ar.RoleARN, func(opts *stscreds.AssumeRoleOptions) {
 		opts.RoleSessionName = ar.SessionName
