@@ -1,22 +1,27 @@
 package awsbase
 
 import (
-	"log"
+	"context"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/iam"
 	"github.com/aws/aws-sdk-go-v2/service/sts"
+	"github.com/hashicorp/aws-sdk-go-base/v2/logging"
 )
 
 // This endpoint resolver is needed when authenticating because the AWS SDK makes internal
 // calls to STS. The resolver should not be attached to the aws.Config returned to the
 // client, since it should configure its own overrides
-func credentialsEndpointResolver(c *Config) aws.EndpointResolverWithOptions {
+func credentialsEndpointResolver(ctx context.Context, c *Config) aws.EndpointResolverWithOptions {
+	logger := logging.RetrieveLogger(ctx)
+
 	resolver := func(service, region string, options ...interface{}) (aws.Endpoint, error) {
 		switch service {
 		case iam.ServiceID:
 			if endpoint := c.IamEndpoint; endpoint != "" {
-				log.Printf("[INFO] Credentials resolution: setting custom IAM endpoint: %s", endpoint)
+				logger.Info(ctx, "Credentials resolution: setting custom IAM endpoint", map[string]any{
+					"tf_aws.iam_client.endpoint": endpoint,
+				})
 				return aws.Endpoint{
 					URL:           endpoint,
 					Source:        aws.EndpointSourceCustom,
@@ -25,12 +30,14 @@ func credentialsEndpointResolver(c *Config) aws.EndpointResolverWithOptions {
 			}
 		case sts.ServiceID:
 			if endpoint := c.StsEndpoint; endpoint != "" {
-				if c.StsRegion != "" {
-					log.Printf("[INFO] Credentials resolution: setting custom STS endpoint: %s (with signing region %s)", endpoint, c.StsRegion)
-					region = c.StsRegion
-				} else {
-					log.Printf("[INFO] Credentials resolution: setting custom STS endpoint: %s", endpoint)
+				fields := map[string]any{
+					"tf_aws.sts_client.endpoint": endpoint,
 				}
+				if c.StsRegion != "" {
+					fields["tf_aws.sts_client.signing_region"] = c.StsRegion
+					region = c.StsRegion
+				}
+				logger.Info(ctx, "Credentials resolution: setting custom STS endpoint", fields)
 				return aws.Endpoint{
 					URL:           endpoint,
 					Source:        aws.EndpointSourceCustom,
