@@ -3,7 +3,6 @@ package awsv1shim
 import ( // nosemgrep: no-sdkv2-imports-in-awsv1shim
 	"context"
 	"fmt"
-	"log"
 	"os"
 
 	awsv2 "github.com/aws/aws-sdk-go-v2/aws"
@@ -15,18 +14,19 @@ import ( // nosemgrep: no-sdkv2-imports-in-awsv1shim
 	"github.com/hashicorp/aws-sdk-go-base/v2/awsv1shim/v2/tfawserr"
 	"github.com/hashicorp/aws-sdk-go-base/v2/internal/awsconfig"
 	"github.com/hashicorp/aws-sdk-go-base/v2/internal/constants"
+	"github.com/hashicorp/aws-sdk-go-base/v2/logging"
 )
 
 // getSessionOptions attempts to return valid AWS Go SDK session authentication
 // options based on pre-existing credential provider, configured profile, or
 // fallback to automatically a determined session via the AWS Go SDK.
-func getSessionOptions(awsC *awsv2.Config, c *awsbase.Config) (*session.Options, error) {
-	useFIPSEndpoint, _, err := awsconfig.ResolveUseFIPSEndpoint(context.Background(), awsC.ConfigSources)
+func getSessionOptions(ctx context.Context, awsC *awsv2.Config, c *awsbase.Config) (*session.Options, error) {
+	useFIPSEndpoint, _, err := awsconfig.ResolveUseFIPSEndpoint(ctx, awsC.ConfigSources)
 	if err != nil {
 		return nil, fmt.Errorf("error resolving FIPS endpoint configuration: %w", err)
 	}
 
-	useDualStackEndpoint, _, err := awsconfig.ResolveUseDualStackEndpoint(context.Background(), awsC.ConfigSources)
+	useDualStackEndpoint, _, err := awsconfig.ResolveUseDualStackEndpoint(ctx, awsC.ConfigSources)
 	if err != nil {
 		return nil, fmt.Errorf("error resolving dual-stack endpoint configuration: %w", err)
 	}
@@ -51,7 +51,7 @@ func getSessionOptions(awsC *awsv2.Config, c *awsbase.Config) (*session.Options,
 	}
 
 	if !c.SuppressDebugLog {
-		options.Config.LogLevel = aws.LogLevel(aws.LogDebugWithHTTPBody | aws.LogDebugWithRequestRetries | aws.LogDebugWithRequestErrors)
+		options.Config.LogLevel = aws.LogLevel(aws.LogOff)
 		options.Config.Logger = debugLogger{}
 	}
 
@@ -63,7 +63,7 @@ func getSessionOptions(awsC *awsv2.Config, c *awsbase.Config) (*session.Options,
 			return nil, err
 		}
 		options.CustomCABundle = reader
-	} else if reader, found, err := resolveCustomCABundle(context.Background(), awsC.ConfigSources); err != nil {
+	} else if reader, found, err := resolveCustomCABundle(ctx, awsC.ConfigSources); err != nil {
 		return nil, fmt.Errorf("error resolving custom CA bundle configuration: %w", err)
 	} else if found {
 		options.CustomCABundle = reader
@@ -72,9 +72,15 @@ func getSessionOptions(awsC *awsv2.Config, c *awsbase.Config) (*session.Options,
 	return options, nil
 }
 
-// GetSession attempts to return valid AWS Go SDK session.
-func GetSession(awsC *awsv2.Config, c *awsbase.Config) (*session.Session, error) {
-	options, err := getSessionOptions(awsC, c)
+const loggerName string = "aws-base-v1"
+
+// GetSession returns an AWS Go SDK session.
+func GetSession(ctx context.Context, awsC *awsv2.Config, c *awsbase.Config) (*session.Session, error) {
+	// var loggerFactory tfLoggerFactory
+	ctx, logger := logging.New(ctx, loggerName)
+	ctx = logging.RegisterLogger(ctx, logger)
+
+	options, err := getSessionOptions(ctx, awsC, c)
 	if err != nil {
 		return nil, err
 	}
@@ -84,7 +90,7 @@ func GetSession(awsC *awsv2.Config, c *awsbase.Config) (*session.Session, error)
 		if tfawserr.ErrCodeEquals(err, "NoCredentialProviders") {
 			return nil, c.NewNoValidCredentialSourcesError(err)
 		}
-		return nil, fmt.Errorf("Error creating AWS session: %w", err)
+		return nil, fmt.Errorf("creating AWS session: %w", err)
 	}
 
 	// Set retries after resolving credentials to prevent retries during resolution
@@ -96,10 +102,18 @@ func GetSession(awsC *awsv2.Config, c *awsbase.Config) (*session.Session, error)
 
 	sess.Handlers.Build.PushBack(userAgentFromContextHandler)
 
+	if !c.SuppressDebugLog {
+		sess.Handlers.Send.PushFrontNamed(requestLogger)
+		sess.Handlers.Send.PushBackNamed(responseLogger)
+	}
+
 	// Add custom input from ENV to the User-Agent request header
 	// Reference: https://github.com/terraform-providers/terraform-provider-aws/issues/9149
 	if v := os.Getenv(constants.AppendUserAgentEnvVar); v != "" {
-		log.Printf("[DEBUG] Using additional User-Agent Info: %s", v)
+		logger.Debug(ctx, "Adding User-Agent info", map[string]any{
+			"source": fmt.Sprintf("envvar(%q)", constants.AppendUserAgentEnvVar),
+			"value":  v,
+		})
 		sess.Handlers.Build.PushBack(request.MakeAddToUserAgentFreeFormHandler(v))
 	}
 
@@ -110,19 +124,25 @@ func GetSession(awsC *awsv2.Config, c *awsbase.Config) (*session.Session, error)
 	// NOTE: This logic can be fooled by other request errors raising the retry count
 	//       before any networking error occurs
 	sess.Handlers.Retry.PushBack(func(r *request.Request) {
+		logger := logging.RetrieveLogger(r.Context())
+
 		if r.RetryCount < constants.MaxNetworkRetryCount {
 			return
 		}
 		// RequestError: send request failed
 		// caused by: Post https://FQDN/: dial tcp: lookup FQDN: no such host
 		if tfawserr.ErrMessageAndOrigErrContain(r.Error, request.ErrCodeRequestError, "send request failed", "no such host") {
-			log.Printf("[WARN] Disabling retries after next request due to networking issue")
+			logger.Warn(ctx, "Disabling retries after next request due to networking error", map[string]any{
+				"error": r.Error,
+			})
 			r.Retryable = aws.Bool(false)
 		}
 		// RequestError: send request failed
 		// caused by: Post https://FQDN/: dial tcp IPADDRESS:443: connect: connection refused
 		if tfawserr.ErrMessageAndOrigErrContain(r.Error, request.ErrCodeRequestError, "send request failed", "connection refused") {
-			log.Printf("[WARN] Disabling retries after next request due to networking issue")
+			logger.Warn(ctx, "Disabling retries after next request due to networking error", map[string]any{
+				"error": r.Error,
+			})
 			r.Retryable = aws.Bool(false)
 		}
 	})
